Implement mount promises.
This change implements a new gVisor-specific feature called mount promises.
You can mark a VirtualDentry as a mount promise with RegisterMountPromise.
After that, any VFS operation (open, link, mknod, etc) on that vd will block
the calling process and emit a SentryMountPromiseBlockEvent.

VFS blocks until the vd is either mounted over, or the mount promise timeout
elapses. After that the VFS operation proceeds as normal.

This feature is useful for any external process that needs to be available
when a user accesses a certain filesystem, but not necessarily during startup.
A FUSE server is one example of this kind of external process.

PiperOrigin-RevId: 521002015