`runsc metric-server`: Add metric for interesting traits of sandbox specs.
This is useful to export data about a sandbox that isn't gVisor configuration
details (that would be the role of the `sandbox_metadata`) metric.
`sandbox_spec` exports data about the specs of containers within a sandbox.

Currently that is the following bits of information:

- Does any of the container run as UID 0? Useful to check whether it has all
  capabilities.
- What OCI runtime version do the containers use?

Other interesting details might be number of containers, number of read-only
vs read-write mounts, and so on. But those might be better off as separate
metrics since they are numerical and could even be per-container rather than
per-pod.

PiperOrigin-RevId: 521036962